Added scene parser test.
* Add more test data.

* Add scene parsing test with real data.

* Add script to download test data.

* Fixed cloud being mutated while reading indices from it.

* Preliminary debugging fixes of scene parser.

* Comment out failing parts of test, fix is a TODO.

+++ rapid_perception/scripts/download_test_data.sh
@@ -4,4 +4,15 @@ if [ ! -e $PKG_PATH/test_data ]
 then
   mkdir $PKG_PATH/test_data
 fi
-wget "https://drive.google.com/uc?export=download&id=0B77PnOCaAq8sNVNHNFpZNnk2aUU" -O `rospack find rapid_perception`/test_data/table4.bag
+if [ ! -e $PKG_PATH/test_data/table0.bag ]
+then
+  wget "https://drive.google.com/uc?export=download&id=0B77PnOCaAq8sWUJNZVB0b1l0cUU" -O `rospack find rapid_perception`/test_data/table0.bag
+fi
+if [ ! -e $PKG_PATH/test_data/table1.bag ]
+then
+  wget "https://drive.google.com/uc?export=download&id=0B77PnOCaAq8sSk1DTW4yQmhHSU0" -O `rospack find rapid_perception`/test_data/table1.bag
+fi
+if [ ! -e $PKG_PATH/test_data/table4.bag ]
+then
+  wget "https://drive.google.com/uc?export=download&id=0B77PnOCaAq8sNVNHNFpZNnk2aUU" -O `rospack find rapid_perception`/test_data/table4.bag
+fi
